[FIR] Warn when using an intersection type for reified type parameter
When an intersection type is inferred and used for a reified type
parameter, subtle bugs may occur because the reified type resolves to
the upper bound of the intersection (often Any?). This may cause
unexpected behavior if the type is used for instantiating a class like
in Kotlin serialization.

Instead of having unpredictable behavior at runtime, issue a warning on
the type parameter, so the user may explicitly specify the type desired.
This warning may be upgraded to an error in a future release: KT-71420.

^KT-52469 Fixed