KT-70920: Add support for nullable primitive types
Implemented handling for nullable primitive types in Swift Export.
The nuance is
that an optional primitive type gets boxed on the kotlin side,
making the handling of them harder in the compiler bridges.
This is solved by using NSNumber on the C side to represent such types.
To overcome the pitfalls of Objective-C export,
generated Swift code automatically unwraps NSNumber into the expected Swift
primitive.

^KT-70920 fixed